+++ src/spec/ruby/jruby/rack/integration_spec.rb
@@ -87,14 +87,16 @@ describe "integration" do
 
     let(:servlet_context) { new_servlet_context(base_path) }
 
-    it "initializes (pooling by default)" do
+    it "initializes pooling when min/max set" do
+      servlet_context.addInitParameter('jruby.min.runtimes', '1')
+      servlet_context.addInitParameter('jruby.max.runtimes', '2')
+
       listener = org.jruby.rack.rails.RailsServletContextListener.new
       listener.contextInitialized javax.servlet.ServletContextEvent.new(servlet_context)
 
       rack_factory = servlet_context.getAttribute("rack.factory")
       rack_factory.should be_a(RackApplicationFactory)
       rack_factory.should be_a(PoolingRackApplicationFactory)
-      rack_factory.should respond_to(:realFactory)
       rack_factory.realFactory.should be_a(RailsRackApplicationFactory)
 
       servlet_context.getAttribute("rack.context").should be_a(RackContext)
@@ -103,9 +105,7 @@ describe "integration" do
       rack_factory.getApplication.should be_a(DefaultRackApplication)
     end
 
-    it "initializes threadsafe!" do
-      servlet_context.addInitParameter('jruby.max.runtimes', '1')
-
+    it "initializes shared (thread-safe) by default" do
       listener = org.jruby.rack.rails.RailsServletContextListener.new
       listener.contextInitialized javax.servlet.ServletContextEvent.new(servlet_context)
 
@@ -117,282 +117,18 @@ describe "integration" do
       rack_factory.getApplication.should be_a(DefaultRackApplication)
     end
 
-  end
+    it "initializes shared (thread-safe) whem max runtimes is 1" do
+      servlet_context.addInitParameter('jruby.max.runtimes', '1')
 
-  describe 'rails 3.0', :lib => :rails30 do
+      listener = org.jruby.rack.rails.RailsServletContextListener.new
+      listener.contextInitialized javax.servlet.ServletContextEvent.new(servlet_context)
 
-    before(:all) { copy_gemfile("rails30") }
+      rack_factory = servlet_context.getAttribute("rack.factory")
+      rack_factory.should be_a(RackApplicationFactory)
+      rack_factory.should be_a(SharedRackApplicationFactory)
+      rack_factory.realFactory.should be_a(RailsRackApplicationFactory)
 
-    def base_path; "file://#{STUB_DIR}/rails30" end
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before :all do
-        initialize_rails nil, base_path
-      end
-
-      it "loaded rack ~> 1.2" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.2'
-      end
-
-      it "disables rack's chunked support (by default)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        expect_to_have_monkey_patched_chunked
-      end
-
-    end
-
-    context "initialized (custom)" do
-
-      before :all do
-        initialize_rails 'custom', base_path
-      end
-
-      it "booted a custom env with a custom logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_eql_to "Rails.env", 'custom'
-        should_eval_as_not_nil "Rails.logger"
-        should_eval_as_eql_to "Rails.logger.class.name", 'CustomLogger'
-      end
-
-    end
-
-  end
-
-  describe 'rails 3.1', :lib => :rails31 do
-
-    before(:all) { copy_gemfile("rails31") }
-
-    def base_path; "file://#{STUB_DIR}/rails31" end
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before :all do
-        initialize_rails 'production', base_path
-      end
-      after(:all)  { restore_rails }
-
-      it "loaded META-INF/init.rb" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(WARBLER_CONFIG)"
-      end
-
-      it "loaded rack ~> 1.3" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.3'
-      end
-
-      it "booted with a servlet logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_not_nil "Rails.logger"
-        should_eval_as_not_nil "Rails.logger.instance_variable_get(:'@logdev')" # Logger::LogDevice
-        should_eval_as_eql_to "Rails.logger.instance_variable_get(:'@logdev').dev.class.name", 'JRuby::Rack::ServletLog'
-
-        should_eval_as_eql_to "Rails.logger.level", Logger::DEBUG
-      end
-
-      it "disables rack's chunked support (by default)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        expect_to_have_monkey_patched_chunked
-      end
-
-    end
-
-  end
-
-  describe 'rails 3.2', :lib => :rails32 do
-
-    before(:all) { copy_gemfile("rails32") }
-
-    def base_path; "file://#{STUB_DIR}/rails32" end
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before(:all) { initialize_rails 'production', base_path }
-      after(:all)  { restore_rails }
-
-      it "loaded rack ~> 1.4" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.4'
-      end
-
-      it "booted with a servlet logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_not_nil "Rails.logger"
-        should_eval_as_eql_to "Rails.logger.class.name", 'ActiveSupport::TaggedLogging'
-        should_eval_as_not_nil "Rails.logger.instance_variable_get(:'@logger')"
-        should_eval_as_eql_to "logger = Rails.logger.instance_variable_get(:'@logger'); " +
-          "logger.instance_variable_get(:'@logdev').dev.class.name", 'JRuby::Rack::ServletLog'
-
-        should_eval_as_eql_to "Rails.logger.level", Logger::INFO
-      end
-
-      it "sets up public_path (as for a war)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_eql_to "Rails.public_path", "#{STUB_DIR}/rails32"
-        # make sure it was set early on (before initializers run) :
-        should_eval_as_not_nil "defined? Rails32::Application::PUBLIC_PATH"
-        should_eval_as_eql_to "Rails32::Application::PUBLIC_PATH", "#{STUB_DIR}/rails32"
-        # check if image_tag resolves path to images correctly :
-        should_eval_as_eql_to %q{
-          config = ActionController::Base.config;
-          asset_paths = ActionView::Helpers::AssetTagHelper::AssetPaths.new(config);
-          image_path = asset_paths.compute_public_path('image.jpg', 'images');
-          image_path[0, 18]
-        }, '/images/image.jpg?'
-      end
-
-      it "disables rack's chunked support (by default)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        expect_to_have_monkey_patched_chunked
-      end
-
-    end
-
-  end
-
-  describe 'rails 4.0', :lib => :rails40 do
-
-    before(:all) { copy_gemfile("rails40") }
-
-    def base_path; "file://#{STUB_DIR}/rails40" end
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before(:all) { initialize_rails 'production', base_path }
-      after(:all)  { restore_rails }
-
-      it "loaded rack ~> 1.5" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.5'
-      end
-
-      it "booted with a servlet logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_not_nil "Rails.logger"
-        # NOTE: TaggedLogging is a module that extends the instance now :
-        should_eval_as_eql_to "Rails.logger.is_a? ActiveSupport::TaggedLogging", true
-        should_eval_as_eql_to "Rails.logger.instance_variable_get(:'@logdev').dev.class.name",
-                              'JRuby::Rack::ServletLog'
-        should_eval_as_eql_to "Rails.logger.level", Logger::INFO
-      end
-
-      it "sets up public_path (as for a war)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_eql_to "Rails.public_path.to_s", "#{STUB_DIR}/rails40" # Pathname
-        # due config.assets.digest = true and since we're asset pre-compiled :
-        #should_eval_as_eql_to %q{
-        #  ActionController::Base.helpers.image_path('image.jpg')[0, 14];
-        #}, '/assets/image-'
-      end
-
-    end
-
-  end
-
-  describe 'rails 4.1', :lib => :rails41 do
-
-    before(:all) do name = :rails41 # copy_gemfile :
-      FileUtils.cp File.join(GEMFILES_DIR, "#{name}.gemfile"), File.join(STUB_DIR, "#{name}/Gemfile")
-      FileUtils.cp File.join(GEMFILES_DIR, "#{name}.gemfile.lock"), File.join(STUB_DIR, "#{name}/Gemfile.lock")
-      Dir.chdir File.join(STUB_DIR, name.to_s)
-    end
-
-    def prepare_servlet_context(servlet_context)
-      servlet_context.addInitParameter('rails.root', "#{STUB_DIR}/rails41")
-      servlet_context.addInitParameter('jruby.rack.layout_class', 'FileSystemLayout')
-    end
-
-    def base_path; "file://#{STUB_DIR}/rails41" end
-    # let(:base_path) { "file://#{STUB_DIR}/rails41" }
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before(:all) { initialize_rails 'production', base_path }
-      after(:all)  { restore_rails }
-
-      it "loaded rack ~> 1.5" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.5'
-      end
-
-      it "booted with a servlet logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_not_nil "Rails.logger"
-        # NOTE: TaggedLogging is a module that extends the instance now :
-        should_eval_as_eql_to "Rails.logger.is_a? ActiveSupport::TaggedLogging", true
-        should_eval_as_eql_to "Rails.logger.instance_variable_get(:'@logdev').dev.class.name",
-                              'JRuby::Rack::ServletLog'
-        should_eval_as_eql_to "Rails.logger.level", Logger::INFO
-      end
-
-      it "sets up public_path (as for a war)" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_eql_to "Rails.public_path.to_s", "#{STUB_DIR}/rails41/public"
-      end
-
-    end
-
-  end
-
-  describe 'rails 2.3', :lib => :rails23 do
-
-    before(:all) do
-      copy_gemfile('rails23')
-      path = File.join(STUB_DIR, 'rails23/WEB-INF/init.rb') # hard-coded RAILS_GEM_VERSION
-      File.open(path, 'w') { |f| f << "RAILS_GEM_VERSION = '#{Rails::VERSION::STRING}'\n" }
-    end
-
-    let(:base_path) { "file://#{STUB_DIR}/rails23" }
-
-    it_should_behave_like 'a rails app'
-
-    context "initialized" do
-
-      before(:all) { initialize_rails 'production', base_path }
-      after(:all)  { restore_rails }
-
-      it "loaded rack ~> 1.1" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rack.release)"
-        should_eval_as_eql_to "Rack.release.to_s[0, 3]", '1.1'
-      end
-
-      it "booted with a servlet logger" do
-        @runtime = @rack_factory.getApplication.getRuntime
-        should_eval_as_not_nil "defined?(Rails)"
-        should_eval_as_not_nil "defined?(Rails.logger)"
-
-        should_eval_as_not_nil "defined?(ActiveSupport::BufferedLogger) && Rails.logger.is_a?(ActiveSupport::BufferedLogger)"
-        should_eval_as_not_nil "Rails.logger.send(:instance_variable_get, '@log')"
-        should_eval_as_eql_to "log = Rails.logger.send(:instance_variable_get, '@log');" +
-                              "log.class.name", 'JRuby::Rack::ServletLog'
-        should_eval_as_eql_to "Rails.logger.level", Logger::INFO
-
-        @runtime.evalScriptlet "Rails.logger.debug 'logging works'"
-      end
-
+      rack_factory.getApplication.should be_a(DefaultRackApplication)
     end
 
   end